moved hadicap logic to traits, added icons in datatable, added handicap controller

+++ database/migrations/2023_10_28_024308_create_rounds_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('rounds', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('golfer_id');
-            $table->integer('score')->nullable();
+            $table->decimal('score')->nullable();
             $table->string('course_name')->nullable();
             $table->timestamp('date_of_round')->nullable();
         });
